api/requirements.txt
pymongo==4.10.1
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1